Convert package to noarch Python
As this is a pure Python package with no conditional constraints, it is
a good candidate for converting to a `noarch: python` package. This
makes it easy to install under various Python environments and cuts down
on the associated maintenance burden. Given this, go ahead and convert
the recipe to `noarch: python`.